+++ changelog.md
@@ -3,6 +3,10 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](	)
 ## - 2017-11-28 - Yudhvir
+### Changed
+- Form placeholder values 
+
+## - 2017-11-28 - Yudhvir
 ### Added
 - check for fcm token if not available